Fix failing docker post process (#131)
* Apply shallow git cloning option to the 'git fetch' steps in the makejdk.sh script (similar to the git clone steps)

* Surrounded getOpenJDKUpdateAndBuildVersion() with the Unix/linux time() to measure the time taken for it to execute

* Added echoes to inform user about git fetch mode (shallow or deep - depending on the value passed via $SHALLOW_CLONE_OPTION)

* Removed duplicate code blocks displaying user info using value from SHALLOW_CLONE_OPTION

* Do not move the artifact if the built via Docker, let the function(s) post Docker execution move it, as it is placed in a different directory (we have a consistency issue here)

* If it is a docker build, move the .tar.gz from the folder in the container into the working folder so Jenkins can pick it up to archive it

* Changed echo text in jtreg_prep.sh script

* Changed the check for USE_DOCKER to check if it does not contain 'true' - safer approach

* Making USE_DOCKER an public (global) variable - so other scripts can see it too

* Merge branch 'master' into fix-failing-docker-post-process

# Conflicts:
#	makejdk.sh

Also additional changes to makejdk.sh and build.sh - detecting if we are running in the Docker environment and performing tasks accordingly

* Moving the OpenJDK tar file into the target folder once its finished building in the Docker container

* Changing the logic to write to the /openjdk/target folder inside the container, instead writing to the /openjdk/ folder in the Container which is the WORKING_DIR on the host

* Changes to the non-Docker workflow -- removed the step to enter a target folder, also moving the j2sdk-image folder to the working directory

* Removed unused CONTAINER_ID from the script file

* Reverted change pointed out by @gdams, renaming j2sdk-image

* Replace occurrences of j2sdk-image with ${OPENJDK_REPO_TAG} in the build.sh script

* Moved the declaration of OPENJDK_REPO_TAG outside the function, towards the top of the script

* Fixed the blanks value in OPENJDK_REPO_TAG by making it into a generic function and called from the common-functions.sh script

* Revert "Changing the logic to write to the /openjdk/target folder inside the container, instead writing to the /openjdk/ folder in the Container which is the WORKING_DIR on the host"

This reverts commit 8da36aa587e32c731ab2e97c316e1cf1919bb643.

* Extract /openjdk/target into its own constant and used across scripts

* - Renamed the function that generates the openjdk git repo tag, and assigned the results of the function call to OPENJDK_REPO_TAG variable
- Applied shellchecks in multiple places

Fixes post-build process of creating the archive and moving it to the working directory (for non-Docker builds)

* Refactored the removing unnecessary files function to remove duplication

+++ sbin/build.sh
@@ -53,9 +53,10 @@ checkIfDockerIsUsedForBuildingOrNot()
   if [[ -f /.dockerenv ]] ; then
     echo "Detected we're in docker"
     WORKING_DIR=/openjdk/build
-    TARGET_DIR=/openjdk/target
+    TARGET_DIR=/openjdk/target/
     OPENJDK_REPO_NAME=/openjdk
     OPENJDK_DIR="$WORKING_DIR/$OPENJDK_REPO_NAME"
+    USE_DOCKER=true
   fi
 
   # E.g. /openjdk/build if you're building in a Docker container
@@ -159,7 +160,7 @@ stepIntoTheWorkingDirectory()
   echo "Should have the source, I'm at $PWD"
 }
 
-runTheOpenJDKConfigureCommandAndUseThePrebuildConfigParams()
+runTheOpenJDKConfigureCommandAndUseThePrebuiltConfigParams()
 {
   cd "$OPENJDK_DIR" || exit
   CONFIGURED_OPENJDK_ALREADY=$(find -name "config.status")
@@ -169,7 +170,7 @@ runTheOpenJDKConfigureCommandAndUseThePrebuildConfigParams()
   else
     CONFIGURE_ARGS="${CONFIGURE_ARGS} ${CONFIGURE_ARGS_FOR_ANY_PLATFORM}"
 
-    echo "Running ./configure with $CONFIGURE_ARGS"
+    echo "Running ./configure with arguments $CONFIGURE_ARGS"
     # Depends upon the configure command being split for multiple args.  Dont quote it.
     # shellcheck disable=SC2086
     bash ./configure $CONFIGURE_ARGS
@@ -235,12 +236,16 @@ printJavaVersionString()
 
 removingUnnecessaryFiles()
 {
-  echo "Removing unneccessary files now..."
-
-  rm -rf cacerts_area
+  echo "Removing unnecessary files now..."
+  
+  OPENJDK_REPO_TAG=$(getFirstTagFromOpenJDKGitRepo)
+  if [ "$USE_DOCKER" != "true" ] ; then
+     rm -rf cacerts_area
+  fi
 
   cd build/*/images || return
 
+  rm -fr "${OPENJDK_REPO_TAG}" || true
   mv j2sdk-image "${OPENJDK_REPO_TAG}"
 
   # Remove files we don't need
@@ -252,25 +257,39 @@ removingUnnecessaryFiles()
 
 createOpenJDKTarArchive()
 {
-  case "${OS_KERNEL_NAME}" in
-    *cygwin*)
-      zip -r -q OpenJDK.zip ./"${OPENJDK_REPO_TAG}"
-      EXT=".zip" ;;
-    aix)
-      GZIP=-9 tar -cf - ./j2sdk-image/ | gzip -c > OpenJDK.tar.gz
-      EXT=".tar.gz" ;;
-    *)
-      GZIP=-9 tar -czf OpenJDK.tar.gz ./"${OPENJDK_REPO_TAG}"
-      EXT=".tar.gz" ;;
-  esac
-  mv "OpenJDK${EXT}" "${TARGET_DIR}"
-  echo "${good}Your final ${EXT} is here at ${PWD}${normal}"
+  echo "Archiving the build OpenJDK image..."
+
+  OPENJDK_REPO_TAG=$(getFirstTagFromOpenJDKGitRepo)
+  echo "OpenJDK repo tag is ${OPENJDK_REPO_TAG}"
+
+  if [ "$USE_DOCKER" == "true" ] ; then
+     GZIP=-9 tar -czf OpenJDK.tar.gz ./"${OPENJDK_REPO_TAG}"
+     EXT=".tar.gz"
+
+     echo "${good}Moving the artifact to ${TARGET_DIR}${normal}"
+     mv "OpenJDK${EXT}" "${TARGET_DIR}"
+  else
+      case "${OS_KERNEL_NAME}" in
+        *cygwin*)
+          zip -r -q OpenJDK.zip ./"${OPENJDK_REPO_TAG}"
+          EXT=".zip" ;;
+        aix)
+          GZIP=-9 tar -cf - ./"${OPENJDK_REPO_TAG}"/ | gzip -c > OpenJDK.tar.gz
+          EXT=".tar.gz" ;;
+        *)
+          GZIP=-9 tar -czf OpenJDK.tar.gz ./"${OPENJDK_REPO_TAG}"
+          EXT=".tar.gz" ;;
+      esac
+      echo "${good}Your final ${EXT} was created at ${PWD}${normal}"
+
+      echo "${good}Moving the artifact to ${TARGET_DIR}${normal}"
+      mv "OpenJDK${EXT}" "${TARGET_DIR}"
+  fi
+
 }
 
-stepIntoTargetDirectoryAndShowCompletionMessage()
+showCompletionMessage()
 {
-  cd "${TARGET_DIR}"  || return
-  ls
   echo "All done!"
 }
 
@@ -280,9 +299,9 @@ createWorkingDirectory
 downloadingRequiredDependencies # This function is in common-functions.sh
 configureCommandParameters
 stepIntoTheWorkingDirectory
-runTheOpenJDKConfigureCommandAndUseThePrebuildConfigParams
+runTheOpenJDKConfigureCommandAndUseThePrebuiltConfigParams
 buildOpenJDK
 printJavaVersionString
 removingUnnecessaryFiles
 createOpenJDKTarArchive
-stepIntoTargetDirectoryAndShowCompletionMessage
+showCompletionMessage
